refactor logic to subquery class
